+++ checkpoint/orbax/checkpoint/experimental/v1/_src/training/preservation_policies.py
@@ -28,6 +28,7 @@ PreserveAll = preservation_policy_lib.PreserveAll
 LatestN = preservation_policy_lib.LatestN
 EveryNSeconds = preservation_policy_lib.EveryNSeconds
 EveryNSteps = preservation_policy_lib.EveryNSteps
+EveryNStepsClosest = preservation_policy_lib.EveryNStepsClosest
 CustomSteps = preservation_policy_lib.CustomSteps
 BestN = preservation_policy_lib.BestN
 LatestDuration = preservation_policy_lib.LatestDuration